+++ src/cli.ts
@@ -46,17 +46,11 @@ if (args.analyze) (async () => {
 	// Fetch language data
 	const root = args.analyze === true ? '.' : args.analyze;
 	const { files, languages, unknown } = await linguist(root, args);
-	// Make file paths relative
-	for (const [file, lang] of Object.entries(files.results)) {
-		const relFile = file.replace(path.resolve().replace(/\\/g, '/'), '.');
-		delete files.results[file];
-		files.results[relFile] = lang;
-	}
 	// Print output
 	if (!args.json) {
 		const sortedEntries = Object.entries(languages.results).map(([lang, data]) => [lang, data.bytes]).sort((a, b) => a[1] < b[1] ? +1 : -1) as [string, number][];
 		const totalBytes = languages.bytes;
-		console.log(`Analysed ${files.bytes} B from ${files.count} files with linguist-js`);
+		console.log(`Analysed ${files.bytes.toLocaleString()} B from ${files.count} files with linguist-js`);
 		console.log(`\n Language analysis results:`);
 		let i = 0;
 		for (const [lang, bytes] of sortedEntries) {